WIP discovery process class
Working on a discovery class, which will start pyre node, name
and send periodic beacons, and will monitor for publishers.

The interface to the discovery service will be to get the dictionary
of ip address to names (latest). Not clear yet how to update and age
out.

For testing initally this is a stand-alone program with debug prints,
but after, it can do logging via the logger package which is supported.